refactor(view): split the two largest views into their sections
