+++ crates/app/src/view/chat.rs
@@ -1,0 +1,287 @@
+//! The chat column: the transcript in front, and everything stacked under it
+//! that the turn in flight puts there — plan, permission, queue, composer.
+
+use crate::{
+    model::session::{ChatSession, PlanStatus},
+    view::{
+        composer,
+        root::{self, Cydonia, Pane},
+        transcript,
+    },
+};
+use bezel::{
+    gpui::{AnyElement, App, Context, FocusHandle, Focusable as _, Window, div, prelude::*, px},
+    motion::{Fade, Painter},
+    theme::Theme,
+    ui::{
+        icons,
+        widgets::{ButtonStyle, Buttons, Content, Scaffolding},
+    },
+};
+use cacp::schema::PermissionOptionKind;
+
+impl Cydonia {
+    pub fn composer_focus_handle(&self, cx: &App) -> FocusHandle {
+        self.composer.focus_handle(cx)
+    }
+
+    /// Run `f` on the active session. Every composer action is this shape:
+    /// the view knows which session is in front, the model owns it.
+    fn with_active(&mut self, cx: &mut Context<Self>, f: impl FnOnce(&mut ChatSession)) {
+        let Some(id) = self.workspace.read(cx).active_id() else {
+            return;
+        };
+        self.workspace
+            .update(cx, |workspace, cx| workspace.with_session(id, cx, f));
+    }
+
+    pub(crate) fn submit(&mut self, text: String, cx: &mut Context<Self>) {
+        self.with_active(cx, |chat| chat.send(text));
+    }
+
+    pub(crate) fn cancel_turn(&mut self, cx: &mut Context<Self>) {
+        self.with_active(cx, |chat| chat.cancel());
+    }
+
+    /// The composer's agent chip. An ACP session is bound to the process that
+    /// serves it, so picking another agent opens a session rather than
+    /// swapping one out from under a transcript.
+    pub(crate) fn pick_agent(&mut self, ix: usize, cx: &mut Context<Self>) {
+        let entry = self.workspace.read(cx).settings.agents.get(ix).cloned();
+        if let Some(entry) = entry {
+            self.show_pane(Pane::Chat, cx);
+            self.workspace
+                .update(cx, |workspace, cx| workspace.new_session(entry, None, cx));
+        }
+    }
+
+    /// What the composer needs from the session it is pointed at: the agent's
+    /// name, its commands, whether a turn is in flight, and the agents it can
+    /// be swapped for.
+    pub(crate) fn sync_composer(&mut self, cx: &mut Context<Self>) {
+        let workspace = self.workspace.read(cx);
+        let agents: Vec<composer::Agent> = workspace
+            .settings
+            .agents
+            .iter()
+            .map(|entry| composer::Agent {
+                name: entry.name.clone().into(),
+                icon: workspace.agent_icon(&entry.name),
+            })
+            .collect();
+        let chat = workspace.active_session();
+        let placeholder = chat.map_or_else(
+            || "message the agent…".to_owned(),
+            |chat| format!("message {}…", chat.entry.name),
+        );
+        let commands = chat.map(|chat| chat.commands.clone()).unwrap_or_default();
+        let streaming = chat.is_some_and(|chat| chat.streaming);
+        let current = chat
+            .map(|chat| chat.entry.name.clone())
+            .and_then(|name| agents.iter().position(|agent| agent.name == name));
+        self.composer.update(cx, |composer, cx| {
+            composer.set_placeholder(&placeholder, cx);
+            composer.set_commands(&commands, cx);
+            composer.set_streaming(streaming, cx);
+            composer.set_agents(&agents, current, cx);
+        });
+    }
+
+    pub(crate) fn chat(
+        &self,
+        window: &mut Window,
+        cx: &mut Context<Self>,
+    ) -> impl IntoElement + use<> {
+        let theme = Theme::of(cx).clone();
+        let open = self.workspace.read(cx).active_project().is_some();
+        // Nothing to send to: archiving closed the connection, so the composer
+        // stack would be a prompt box wired to a dead process.
+        let live = !self
+            .workspace
+            .read(cx)
+            .active_session()
+            .is_some_and(|chat| chat.archive.is_some());
+        let body = if !open {
+            self.no_project(cx)
+        } else {
+            match self.showing(cx) {
+                Pane::Chat => self.conversation(window, cx),
+                Pane::Board => self.board(cx),
+                Pane::Article => match self.article(cx) {
+                    Some(article) => article,
+                    None => self.conversation(window, cx),
+                },
+                Pane::Table => match self.table(cx) {
+                    Some(table) => table,
+                    None => self.conversation(window, cx),
+                },
+            }
+        };
+
+        let card = div()
+            .flex_1()
+            .min_h_0()
+            // With the rail gone the head band above it already clears the
+            // traffic lights, and a margin on top of that doubles the air.
+            .mt(px(if self.sidebar_open {
+                root::SHELL_INSET
+            } else {
+                0.
+            }))
+            .ml(px(root::SHELL_INSET))
+            .mr(px(root::SHELL_INSET))
+            .flex()
+            .flex_col()
+            .rounded(px(Theme::panel_radius()))
+            .bg(theme.surface)
+            .border_1()
+            .border_color(theme.border)
+            .overflow_hidden()
+            .child(body)
+            .when(open && live, |card| {
+                card.child(
+                    div().flex_none().flex().justify_center().child(
+                        div()
+                            .w_full()
+                            .max_w(px(720.))
+                            .px(px(24.))
+                            .pb(px(20.))
+                            .flex()
+                            .flex_col()
+                            .gap(px(8.))
+                            .children(self.plan(cx))
+                            .children(self.permission(cx))
+                            .children(self.queue(cx))
+                            .child(self.composer.clone()),
+                    ),
+                )
+            });
+
+        div()
+            .flex_1()
+            .min_w_0()
+            .flex()
+            .flex_col()
+            .when(!self.sidebar_open, |column| {
+                column.child(self.rail_head(window, cx))
+            })
+            .child(card)
+            .when(open, |column| column.child(self.pane_switch(cx)))
+    }
+
+    /// The session in front, or the invitation to open one.
+    fn conversation(&self, window: &mut Window, cx: &mut Context<Self>) -> AnyElement {
+        match self.workspace.read(cx).active_id() {
+            Some(id) => self
+                .workspace
+                .update(cx, |workspace, cx| match workspace.session(id) {
+                    Some(chat) => transcript::render(chat, window, cx),
+                    None => div().flex_1().into_any_element(),
+                }),
+            None => Theme::of(cx)
+                .empty_state(
+                    icons::CHAT_ROUND_LINE,
+                    "No session",
+                    "⌘N to start one in this project.",
+                )
+                .flex_1()
+                .into_any_element(),
+        }
+    }
+
+    /// The agent's plan, while it still has something left to do.
+    fn plan(&self, cx: &Context<Self>) -> Option<impl IntoElement + use<>> {
+        let theme = Theme::of(cx).clone();
+        let chat = self.workspace.read(cx).active_session()?;
+        if chat.plan.is_empty() || chat.plan.iter().all(|(_, s)| *s == PlanStatus::Done) {
+            return None;
+        }
+        Some(
+            theme
+                .group_box()
+                .mt(px(0.))
+                .px(px(12.))
+                .py(px(8.))
+                .gap(px(4.))
+                .text_size(px(12.))
+                .children(chat.plan.iter().map(|(text, status)| {
+                    let (icon, tone) = match status {
+                        PlanStatus::Done => (icons::CHECK, theme.success),
+                        PlanStatus::Active => (icons::ALT_ARROW_RIGHT, theme.accent),
+                        PlanStatus::Pending => (icons::CHECKLIST, theme.text_faint),
+                    };
+                    div()
+                        .flex()
+                        .flex_row()
+                        .items_start()
+                        .gap(px(8.))
+                        .text_color(theme.text_muted)
+                        .child(icons::icon(icon).size(px(12.)).text_color(tone))
+                        .child(text.clone())
+                })),
+        )
+    }
+
+    /// The agent's tool-authorization request, one button per option.
+    fn permission(&self, cx: &Context<Self>) -> Option<impl IntoElement + use<>> {
+        let theme = Theme::of(cx).clone();
+        let chat = self.workspace.read(cx).active_session()?;
+        let prompt = chat.permission.as_ref()?;
+        let id = chat.id;
+        let painter = Painter::of(cx);
+        Some(
+            theme
+                .group_box()
+                .mt(px(0.))
+                .border_color(theme.accent)
+                .px(px(12.))
+                .py(px(10.))
+                .gap(px(10.))
+                .child(
+                    div()
+                        .text_size(px(13.))
+                        .text_color(theme.text)
+                        .child(prompt.title.clone()),
+                )
+                .child(div().flex().flex_row().gap(px(8.)).children(
+                    prompt.options.iter().enumerate().map(|(ix, option)| {
+                        let style = match option.kind {
+                            PermissionOptionKind::AllowOnce => ButtonStyle::Prominent,
+                            PermissionOptionKind::RejectOnce
+                            | PermissionOptionKind::RejectAlways => ButtonStyle::Destructive,
+                            _ => ButtonStyle::Ghost,
+                        };
+                        let fade = Fade::new(painter, format!("permission-{id}-{ix}"));
+                        let option_id = option.id.clone();
+                        theme
+                            .button(option.name.clone(), style, Some(fade))
+                            .id(("permission", ix))
+                            .on_click(cx.listener(move |this, _, _, cx| {
+                                let option_id = option_id.clone();
+                                this.workspace.update(cx, |workspace, cx| {
+                                    workspace.with_session(id, cx, |chat| {
+                                        chat.respond_permission(option_id);
+                                    });
+                                });
+                            }))
+                    }),
+                )),
+        )
+    }
+
+    /// Prompts waiting for the in-flight turn.
+    fn queue(&self, cx: &Context<Self>) -> Option<impl IntoElement + use<>> {
+        let theme = Theme::of(cx).clone();
+        let chat = self.workspace.read(cx).active_session()?;
+        if chat.queue.is_empty() {
+            return None;
+        }
+        Some(
+            div().flex().flex_row().flex_wrap().gap(px(6.)).children(
+                chat.queue
+                    .iter()
+                    .map(|text| theme.badge(format!("queued · {text}"))),
+            ),
+        )
+    }
+}
